refactor: move domain reconciliation functions to respective packages
- Move ReconcileDotfiles to resources/dotfiles/reconcile.go
- Move ReconcilePackages to resources/packages/reconcile.go
- Keep ReconcileAll in orchestrator as coordination logic
- Update all imports to use domain-specific reconciliation functions
- Update integration tests to use new function locations
- Remove utilities package as functions are now properly organized
- Preserve generic reconciliation abstraction in resources package
- Complete Task 6.3: Simplify Orchestrator Package

The orchestrator now only handles coordination logic while domain-specific
reconciliation lives in appropriate packages. This maintains the reconciliation
abstraction crucial for AI Lab extensibility.
